Move files to src
Move files.php and create Term class
